Filter products by category and all its descendants

+++ src/main/java/com/ambraspace/etprodaja/model/category/CategoryRepository.java
@@ -3,7 +3,7 @@ package com.ambraspace.etprodaja.model.category;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer>, PagingAndSortingRepository<Category, Integer>
+public interface CategoryRepository extends CrudRepository<Category, Long>, PagingAndSortingRepository<Category, Long>
 {
 
 	Iterable<Category> findByRootCategoryIsTrue();
